requirements.txt
astroid==2.3.1
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-summernote==0.8.11.4
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jsonapi==3.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pbr==5.4.4
psycopg2==2.8.3
PyJWT==1.7.1
pylint==2.4.2
pytz==2019.3
python-decouple==3.3
requests==2.22.0
send-email==201212
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.31.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
whitenoise==5.0.1
wrapt==1.11.2
